fix: split useFactoryEventStream tests for lint line limit (mergeability)
