+++ ui/src/features/dashboard/hooks/useFactoryEventStream.test.tsx
@@ -1,8 +1,8 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
+import type { QueryClient } from "@tanstack/react-query";
 import { act, renderHook, waitFor } from "@testing-library/react";
 import type { PropsWithChildren } from "react";
 
-import type { DashboardSnapshot } from "../../../api/dashboard/types";
 import { FACTORY_EVENT_TYPES } from "../../../api/events";
 import { DEFAULT_FACTORY_SESSION_ID } from "../../../api/session-routing";
 import { createReplayHarness } from "../../../testing/replay-harness";
@@ -12,153 +12,66 @@ import {
 } from "../../current-factory-definition/public";
 import { useDashboardSessionStore } from "../state/dashboardSessionStore";
 import { DashboardSessionProvider } from "../session/dashboard-session-provider";
-import {
-  type WorldState,
-  useFactoryTimelineStore,
-} from "../../timeline/state/factoryTimelineStore";
+import { useFactoryTimelineStore } from "../../timeline/state/factoryTimelineStore";
 import {
   createDefaultDashboardStreamState,
   useDashboardStreamStore,
 } from "../state/dashboardStreamStore";
 import { useFactoryEventStream } from "./useFactoryEventStream";
+import {
+  CANONICAL_SELECTED_TICK_EVENTS,
+  createFactoryEventStreamQueryClient,
+  SEEDED_SNAPSHOT,
+  timelineSnapshot,
+} from "./useFactoryEventStream.fixtures";
 
 const replayHarness = createReplayHarness();
 
-const SEEDED_SNAPSHOT: DashboardSnapshot = {
-  factory_state: "IDLE",
-  runtime: {
-    in_flight_dispatch_count: 0,
-    session: {
-      completed_count: 0,
-      dispatched_count: 0,
-      failed_count: 0,
-      has_data: true,
+function seedFactoryEventStreamStores(): void {
+  useDashboardStreamStore.setState({
+    streamState: createDefaultDashboardStreamState(),
+  });
+  useDashboardSessionStore.setState({
+    pausedSessionIDs: [],
+    selectedSessionID: DEFAULT_FACTORY_SESSION_ID,
+  });
+  useFactoryTimelineStore.setState({
+    events: [],
+    latestTick: SEEDED_SNAPSHOT.tick_count,
+    mode: "current",
+    receivedEventIDs: [],
+    selectedTick: SEEDED_SNAPSHOT.tick_count,
+    worldViewCache: {
+      [SEEDED_SNAPSHOT.tick_count]: timelineSnapshot(SEEDED_SNAPSHOT),
     },
-  },
-  tick_count: 3,
-  topology: {
-    edges: [],
-    workstation_node_ids: [],
-    workstation_nodes_by_id: {},
-  },
-  uptime_seconds: 12,
-};
-
-function timelineSnapshot(snapshot: DashboardSnapshot): WorldState {
-  return {
-    ...snapshot,
-    relationsByWorkID: {},
-    tracesByWorkID: {},
-    workstationRequestsByDispatchID: {},
-    workRequestsByID: {},
-  };
+  });
 }
 
-const CANONICAL_SELECTED_TICK_EVENTS = [
-  {
-    context: {
-      eventTime: "2026-04-25T20:00:01Z",
-      sequence: 1,
-      tick: 1,
-    },
-    id: "event-1",
-    payload: {
-      factory: {
-        workTypes: [{
-          name: "story",
-          states: [
-            { name: "new", type: "INITIAL" },
-            { name: "done", type: "TERMINAL" },
-          ],
-        }],
-        workstations: [
-          {
-            id: "review",
-            inputs: [{ state: "new", workType: "story" }],
-            name: "Review",
-            outputs: [{ state: "done", workType: "story" }],
-            worker: "reviewer",
-          },
-        ],
-        workers: [
-          {
-            model: "gpt-5.4",
-            modelProvider: "codex",
-            name: "reviewer",
-            type: "MODEL_WORKER",
-          },
-        ],
-      },
-    },
-    type: FACTORY_EVENT_TYPES.initialStructureRequest,
-  },
-  {
-    context: {
-      eventTime: "2026-04-25T20:00:02Z",
-      requestId: "request-story-1",
-      sequence: 2,
-      tick: 2,
-      traceIds: ["trace-story-1"],
-      workIds: ["work-story-1"],
-    },
-    id: "event-2",
-    payload: {
-      type: "FACTORY_REQUEST_BATCH",
-      works: [
-        {
-          name: "Canonical Story",
-          trace_id: "trace-story-1",
-          work_id: "work-story-1",
-          work_type_name: "story",
-        },
-      ],
-    },
-    type: FACTORY_EVENT_TYPES.workRequest,
-  },
-];
+function resetFactoryEventStreamStores(): void {
+  replayHarness.reset();
+  useDashboardStreamStore.setState({
+    streamState: createDefaultDashboardStreamState(),
+  });
+  useDashboardSessionStore.setState({
+    pausedSessionIDs: [],
+    selectedSessionID: DEFAULT_FACTORY_SESSION_ID,
+  });
+  useFactoryTimelineStore.getState().reset();
+}
 
-describe("useFactoryEventStream", () => {
-  let queryClient: QueryClient;
+describe("useFactoryEventStream transport", () => {
+  let queryClient = createFactoryEventStreamQueryClient();
   const receivedEvents: unknown[] = [];
 
   beforeEach(() => {
     replayHarness.install();
     receivedEvents.length = 0;
-    queryClient = new QueryClient({
-      defaultOptions: {
-        mutations: { retry: false },
-        queries: { retry: false },
-      },
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useDashboardSessionStore.setState({
-      pausedSessionIDs: [],
-      selectedSessionID: DEFAULT_FACTORY_SESSION_ID,
-    });
-    useFactoryTimelineStore.setState({
-      events: [],
-      latestTick: SEEDED_SNAPSHOT.tick_count,
-      mode: "current",
-      receivedEventIDs: [],
-      selectedTick: SEEDED_SNAPSHOT.tick_count,
-      worldViewCache: {
-        [SEEDED_SNAPSHOT.tick_count]: timelineSnapshot(SEEDED_SNAPSHOT),
-      },
-    });
+    queryClient = createFactoryEventStreamQueryClient();
+    seedFactoryEventStreamStores();
   });
 
   afterEach(() => {
-    replayHarness.reset();
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useDashboardSessionStore.setState({
-      pausedSessionIDs: [],
-      selectedSessionID: DEFAULT_FACTORY_SESSION_ID,
-    });
-    useFactoryTimelineStore.getState().reset();
+    resetFactoryEventStreamStores();
   });
 
   it("opens the session-scoped events stream and delivers compacted events to onEvent", async () => {
@@ -271,6 +184,34 @@ describe("useFactoryEventStream", () => {
     await waitFor(() => {
       expect(replayHarness.getStreams()).toHaveLength(2);
     });
+  });
+
+  it("does not open a stream when sessionID is null", () => {
+    renderHook(
+      () =>
+        useFactoryEventStream({
+          enabled: false,
+          onEvent: () => {},
+          sessionID: null,
+        }),
+      { wrapper: createWrapper(queryClient) },
+    );
+
+    expect(replayHarness.getStreams()).toHaveLength(0);
+  });
+});
+
+describe("useFactoryEventStream side effects", () => {
+  let queryClient = createFactoryEventStreamQueryClient();
+
+  beforeEach(() => {
+    replayHarness.install();
+    queryClient = createFactoryEventStreamQueryClient();
+    seedFactoryEventStreamStores();
+  });
+
+  afterEach(() => {
+    resetFactoryEventStreamStores();
   });
 
   it("updates factory definition queries from FACTORY_CHANGE events", async () => {
@@ -386,20 +327,6 @@ describe("useFactoryEventStream", () => {
         message: "Factory event stream disconnected. Showing last event state.",
       });
     });
-  });
-
-  it("does not open a stream when sessionID is null", () => {
-    renderHook(
-      () =>
-        useFactoryEventStream({
-          enabled: false,
-          onEvent: () => {},
-          sessionID: null,
-        }),
-      { wrapper: createWrapper(queryClient) },
-    );
-
-    expect(replayHarness.getStreams()).toHaveLength(0);
   });
 });
